moved the analytics and utilities
