+++ app/analytics/user_analytics.py
